Исправление описания миграций. Правки значений по умолчанию.

+++ migrations/m190101_100008_update8.php
@@ -34,10 +34,18 @@ class m190101_100008_update8 extends Migration
 			$this->createIndex('{{%idx-users-org_id}}', '{{%users}}', 'org_id');
 		}
 		
+		if (!isset($table->columns['employ_date'])) {
+			//добавляем дату приема
+			$this->addColumn('{{%users}}', 'employ_date', $this->string(16)->Null()->comment('Дата приема')->after('manager_id')->append('COLLATE utf8mb4_unicode_ci'));
+		}
+		if (!isset($table->columns['resign_date'])) {
+			//добавляем дату приема
+			$this->addColumn('{{%users}}', 'resign_date', $this->string(16)->Null()->comment('Дата увольнения')->after('employ_date')->append('COLLATE utf8mb4_unicode_ci'));
+		}
+		
 		$table=$this->db->getTableSchema('{{%login_journal}}');
 		if (isset($table->foreignKeys['login_journal_ibfk_1'])) $this->dropForeignKey('{{%login_journal_ibfk_1}}','{{%login_journal}}');
 		if (isset($table->foreignKeys['login_journal_ibfk_2'])) $this->dropForeignKey('{{%login_journal_ibfk_2}}','{{%login_journal}}');
-		
 		
 		$sql=<<<SQL
 			SET FOREIGN_KEY_CHECKS=0;
@@ -140,11 +148,24 @@ SQL;
 		$this->execute($sql);
 
 		//вот это вот тоже хз как пройдет если уже есть совпадения между организациями
-		$this->dropPrimaryKey('PRIMARY','org_struct');
-		$this->addPrimaryKey('PRIMARY','org_struct',['id']);
+		$table=$this->db->getTableSchema('{{%org_struct}}');
 		
-		$this->dropColumn('users','org_id');
-		$this->dropColumn('users','employee_id');
+		if (isset($table->columns['org_id'])) {
+			$this->dropPrimaryKey('PRIMARY', 'org_struct');
+			$this->addPrimaryKey('', 'org_struct', ['id']);
+		}
+		
+		$table=$this->db->getTableSchema('{{%users}}');
+		
+		if (isset($table->columns['org_id']))
+			$this->dropColumn('users', 'org_id');
+		if (isset($table->columns['employee_id']))
+			$this->dropColumn('users', 'employee_id');
+		if (isset($table->columns['employ_date']))
+			$this->dropColumn('users', 'employ_date');
+		if (isset($table->columns['resign_date']))
+			$this->dropColumn('users', 'resign_date');
+		
 
 		$this->dropColumn('org_struct','org_id');
 	}
